Avoid repeating a URL in the HTML to text conversion
When you type a URL into the WYSIWYG editor, it gets converted
to a link automatically. However, when it is then converted to
plain text, htmltoText produced the URL twice: once as the 'link
text' and once as the link itself.

This change makes sure that if the link text is identical to the
link itself, it is not repeated.