fix comment update mechanism
This fixes a bug that prevents users from updating comments due to the following error:

    Uncaught TypeError: Cannot read properties of undefined (reading 'contains')
        at newCommentBox (comment.js:145:27)
        at editPost (comment.js:131:14)
        at HTMLButtonElement.onclick (<redacted>:2008:77)

The fix adds an ID (the comment slug value) to the div container of each comment. Now, editPost
will use the slug value to identify the div container and infer child nodes from that, rather than
using a reference to the dt node and inferring the relevant nodes from that dt node.