feat: Added CPEA manual-entry functionality to cve5 editor.
This adds support for manual entry of CPE Applicability statements
when entering CVE5 data. This is done as an alternative to the
auto-generation of CPE Applicability statements already present
in the editor. The user controls which of the two options
applies. The automated entry only applies if the user fills out
the manual entry form field and submits it, and stops applying
if the user resets the field.

This feature serves two purposes:

- Enables easy copying over of CPE Applicability information
  already written that uses the NVD format, which CVE largely
  adopted when incorporated support for CPE Applicability
  statements.
- Enables entry of more complex CPE Applicability statements
  than what can be supported via auto-generation from the
  "affected" array alone.

Helped-by: Andrew Lilley Brinker <abrinker@mitre.org>
Signed-off-by: Devin Lake <dlake@mitre.org>